requirements.txt
addict==2.4.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
contourpy==1.2.1
cycler==0.12.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
fastjsonschema==2.20.0
Flask==3.0.3
fonttools==4.53.0
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nptyping==2.5.0
numpy==1.26.4
open3d==0.18.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
pynrrd==1.0.0
pyparsing==3.1.2
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rpds-py==0.19.0
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
SimpleITK==2.3.1
six==1.16.0
-e git+https://github.com/Eguchi-Ryo/3d_registratoion_from_3cs.git@962d38ccde995efeb7347e41c9862898df0854a4#egg=skelton_registration_mac_dev03
tenacity==8.5.0
threadpoolctl==3.5.0
tifffile==2024.8.10
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
